Update Windows CI to windows-2025
windows-2019 will be fully retired by the end of June, and 2025 is the latest version.

This means that Windows build artifacts will be built using MSVC 2022, not MSVC 2019. They should be ABI-compatible across this change, since they expose no third-party types and Microsoft say that their compiler and standard library didn't break ABI between 2019 and 2022, but a newer version of the redistributable might be needed.